+++ src/Core/AmazonService.php
@@ -26,6 +26,9 @@ use OxidSolutionCatalysts\AmazonPay\Model\Order as AmazonOrder;
 use Psr\Log\LoggerInterface;
 use stdClass;
 
+/**
+ * @SuppressWarnings(PHPMD.ExcessiveClassLength)
+ */
 class AmazonService
 {
     /**
@@ -371,7 +374,10 @@ class AmazonService
         $response = PhpHelper::jsonToArray($result['response']);
 
         // in case of error, the resulting structure is different...
-        if (!isset($result['response'], $result['status']) || $result['status'] !== 200) {
+        if (
+            !isset($result['response'], $result['status']) ||
+            ($result['status'] !== 200 && $result['status'] !== 202)
+        ) {
             $this->showErrorOnRedirect($logger, $result, (string)$basket->getOrderId());
         }
 
@@ -453,7 +459,12 @@ class AmazonService
         /** @var string $orderCurrencyName */
         $orderCurrencyName = $order->getOrderCurrency()->name;
 
-        if ($refundAmount < 0 || $refundAmount > $this->getMaximalRefundAmount($orderId)) {
+        // amounts needs to be cast with same precision level or
+        // else even if numbers looks the same the compare will be wrong
+        if (
+            $refundAmount < 0 ||
+            round($refundAmount, 2) > round($this->getMaximalRefundAmount($orderId), 2)
+        ) {
             Registry::getUtilsView()->addErrorToDisplay(
                 Registry::getLang()->translateString(
                     "OSC_AMAZONPAY_REFUND_ANNOTATION"
@@ -609,13 +620,24 @@ class AmazonService
         if ($order->load($orderId)) {
             switch ($response['statusDetails']['state']) {
                 case "Declined":
-                    $order->updateAmazonPayOrderStatus('AMZ_AUTH_OR_CAPT_DECLINED', $result);
+                    $order->updateAmazonPayOrderStatus('AMZ_AUTH_OR_CAPT_DECLINED', [
+                        'result' => $result
+                    ]);
                     break;
                 case "Pending":
                     $order->updateAmazonPayOrderStatus('AMZ_PAYMENT_PENDING', $result);
                     break;
+                case 'Authorized':
+                    $order->updateAmazonPayOrderStatus('AMZ_2STEP_AUTH_OK', [
+                        'chargeAmount' => $response['chargeAmount']['amount'],
+                        'chargeId'     => $response['chargeId']
+                    ]);
+                    break;
                 case "Captured":
-                    $order->updateAmazonPayOrderStatus('AMZ_AUTH_AND_CAPT_OK', $result);
+                    $order->updateAmazonPayOrderStatus('AMZ_AUTH_AND_CAPT_OK', [
+                        'chargeAmount' => $response['chargeAmount']['amount'],
+                        'chargeId'     => $response['chargeId']
+                    ]);
                     break;
             }
         }